# Implement a faster default reader