Hello! It's me again!

In pursuit of more performance gains, I wonder if you'd be interested in this modification. In my last PR I made it possible to swap out the file reader implementation. This change adds a faster default reader.

`File.binread` reads the entire file into memory. However, we don't have to do that if we use `seek` and `read`. Also, according to the MaxMindDB spec, the metadata section can only take up to 128kb. This means that we only have to read, at most, 128kb into memory to fully parse the metadata. 

The speedup is not quite as dramatic as using `Mmap`, but it is still significant. 

Running the tests against `master`:
```
Finished in 0.82818 seconds (files took 0.15708 seconds to load)
110 examples, 0 failures
```

Running the tests against this branch:
```
Finished in 0.13147 seconds (files took 0.16954 seconds to load)
110 examples, 0 failures
```

Around a ~6x increase.