@yhirose 2018-06-23
@aaron-stripe, thanks for another contribution! I merged it, and bumped up npm version to 0.1.19. Good job!