@damonmorgan 2018-07-20
I upgraded and noticed performance degradation so ran some tests. It looks like `File.binread(path)` is slower to initialize, but the new `MaxMindDB::Reader.new(path)` is slower per call, which is why it appears faster when running the specs because each test initializes the reader each time.

But it real world usage, I would expect initializing the db once and then running many lookups against it?

```
require 'maxminddb'
require "benchmark"

def perform(db)
  10000.times { db.lookup("%d.%d.%d.%d" % [rand(256), rand(256), rand(256), rand(256)]) }
end

MaxMindDB::DEFAULT_FILE_READER = proc { |path| File.binread(path) }
bin_read = MaxMindDB.new('spec/cache/GeoLite2-Country.mmdb')

MaxMindDB::DEFAULT_FILE_READER = proc { |path| MaxMindDB::Reader.new(path) }
reader = MaxMindDB.new('spec/cache/GeoLite2-Country.mmdb')

Benchmark.bm do |b|
  b.report("binread") { perform(bin_read) }
  b.report("reader") { perform(reader) }
end

>>
       user     system      total        real
binread  2.186050   0.004479   2.190529 (  2.201890)
reader 11.038942   3.796670  14.835612 ( 14.902528)
```
And when used in a forking environment (puma, unicorn) it creates contention issues

```
require 'maxminddb'

def forking(db)
  3.times do
    fork do
      perform(db)
    end
  end

  Process.waitall
end

def perform(db)
  10.times { db.lookup("%d.%d.%d.%d" % [rand(256), rand(256), rand(256), rand(256)]) }
end

MaxMindDB::DEFAULT_FILE_READER = proc { |path| File.binread(path) }
bin_read = MaxMindDB.new('spec/cache/GeoLite2-Country.mmdb')

MaxMindDB::DEFAULT_FILE_READER = proc { |path| MaxMindDB::Reader.new(path) }
reader = MaxMindDB.new('spec/cache/GeoLite2-Country.mmdb')

puts "Start bin_read"
forking(bin_read)
puts "End bin_read"
puts "Start reader"
forking(reader)
puts "End reader"

>>
Start bin_read
End bin_read
Start reader
gems/maxminddb-0.1.20/lib/maxminddb.rb:56:in `block in lookup': invalid file format (RuntimeError)
gems/maxminddb-0.1.20/lib/maxminddb.rb:138:in `decode': undefined method `-' for false:FalseClass (NoMethodError)
gems/maxminddb-0.1.20/lib/maxminddb.rb:85:in `decode': undefined method `ord' for nil:NilClass (NoMethodError)
End reader
```