@aaron-stripe 2018-07-20
Thanks @damonmorgan. I wonder if this is related to https://github.com/yhirose/maxminddb/issues/42.

Sorry for the trouble here! I wonder if we should revert my change while I dig into what's going on here?