@aaron-stripe 2018-07-20
I opened a revert PR here: https://github.com/yhirose/maxminddb/pull/43