@aaron-stripe 2018-07-20
Thinking about this a bit more, I think the ideal solution is to leave `File.binread` as the default reader, and allow users to opt-in to the low-memory reader. As you say, the low-memory reader is thread-safe, but not necessarily process safe.

I can work on a PR that maintains the `File.binread` reader as the default, but keeps the low-memory version around for users who prioritize that behavior.