@damonmorgan 2018-07-20
It will also need to the proviso that it is slower if you follow the common use case as suggested in https://github.com/yhirose/maxminddb/issues/17#issuecomment-139702365