@aaron-stripe 2018-07-20
@damonmorgan would be great to get your feedback on https://github.com/yhirose/maxminddb/pull/44 if you have time.